+++ src/MicroElements.Metadata/Metadata/IMetadataProvider.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
 namespace MicroElements.Metadata
@@ -14,6 +15,22 @@ namespace MicroElements.Metadata
         /// Gets metadata for current instance.
         /// </summary>
         IPropertyContainer Metadata => this.GetInstanceMetadata();
+    }
+
+    /// <summary>
+    /// MetadataProvider statics.
+    /// </summary>
+    public static class MetadataProvider
+    {
+        /// <summary>
+        /// Default property comparer for metadata search.
+        /// </summary>
+        public static readonly IEqualityComparer<IProperty> DefaulPropertyComparer = PropertyComparer.ByTypeAndNameComparer;
+
+        /// <summary>
+        /// Default search options for metadata providers.
+        /// </summary>
+        public static readonly SearchOptions DefaulSearchOptions = SearchOptions.Default.WithPropertyComparer(DefaulPropertyComparer);
     }
 
     /// <summary>
@@ -36,7 +53,7 @@ namespace MicroElements.Metadata
 
             if (!MetadataCache.TryGetValue(instance, out IPropertyContainer propertyList))
             {
-                propertyList = new MutablePropertyContainer();
+                propertyList = new MutablePropertyContainer(searchOptions: MetadataProvider.DefaulSearchOptions);
                 MetadataCache.Add(instance, propertyList);
             }