+++ src/MicroElements.Metadata/Metadata/MetadataProviderExtensions.cs
@@ -18,17 +18,13 @@ namespace MicroElements.Metadata
         /// <param name="metadataProvider">Metadata provider.</param>
         /// <param name="metadataName">Optional metadata name.</param>
         /// <param name="defaultValue">Default value to return if not metadata found.</param>
-        /// <param name="ignoreCase">Use ignore case comparison.</param>
-        /// <param name="searchInParent">Search property in parent if not found in current.</param>
         /// <returns>Metadata or default value if not found.</returns>
         public static TMetadata GetMetadata<TMetadata>(
             this IMetadataProvider metadataProvider,
             string metadataName = null,
-            TMetadata defaultValue = default,
-            bool ignoreCase = true,
-            bool searchInParent = true)
+            TMetadata defaultValue = default)
         {
-            return GetMetadataAsOption<TMetadata>(metadataProvider, metadataName, ignoreCase, searchInParent)
+            return GetMetadataAsOption<TMetadata>(metadataProvider, metadataName)
                 .MatchUnsafe(metadata => metadata, defaultValue);
         }
 
@@ -38,14 +34,10 @@ namespace MicroElements.Metadata
         /// <typeparam name="TMetadata">Metadata type.</typeparam>
         /// <param name="metadataProvider">Metadata provider.</param>
         /// <param name="metadataName">Optional metadata name.</param>
-        /// <param name="ignoreCase">Use ignore case comparison.</param>
-        /// <param name="searchInParent">Search property in parent if not found in current.</param>
         /// <returns>Metadata or default value if not found.</returns>
         public static Option<TMetadata> GetMetadataAsOption<TMetadata>(
             this IMetadataProvider metadataProvider,
-            string metadataName = null,
-            bool ignoreCase = true,
-            bool searchInParent = true)
+            string metadataName = null)
         {
             metadataProvider.AssertArgumentNotNull(nameof(metadataProvider));
 
@@ -53,8 +45,7 @@ namespace MicroElements.Metadata
             var metadata = metadataProvider.Metadata ?? metadataProvider.GetInstanceMetadata();
 
             var propertyValue = metadata.GetPropertyValue<TMetadata>(Search
-                .ByNameOrAlias<TMetadata>(metadataName, ignoreCase)
-                .SearchInParent(searchInParent)
+                .ByNameAndComparer<TMetadata>(metadataName, MetadataProvider.DefaulPropertyComparer)
                 .ReturnNull());
 
             if (propertyValue.HasValue() && !propertyValue.Value.IsNull())
@@ -114,13 +105,11 @@ namespace MicroElements.Metadata
         public static TMetadataProvider ConfigureMetadata<TMetadataProvider, TMetadata>(
             this TMetadataProvider metadataProvider,
             Action<TMetadata> configureMetadata,
-            string metadataName = null,
-            bool ignoreCase = true,
-            bool searchInParent = true)
+            string metadataName = null)
             where TMetadataProvider : IMetadataProvider
             where TMetadata : new()
         {
-            ConfigureMetadata<TMetadata>(metadataProvider, configureMetadata, metadataName, ignoreCase, searchInParent);
+            ConfigureMetadata<TMetadata>(metadataProvider, configureMetadata, metadataName);
 
             return metadataProvider;
         }
@@ -133,20 +122,16 @@ namespace MicroElements.Metadata
         /// <param name="metadataProvider">Target metadata provider.</param>
         /// <param name="configureMetadata">Configure action.</param>
         /// <param name="metadataName">Optional metadata name.</param>
-        /// <param name="ignoreCase">Use ignore case comparison.</param>
-        /// <param name="searchInParent">Search property in parent if not found in current.</param>
         /// <returns>The same metadataProvider.</returns>
         public static IMetadataProvider ConfigureMetadata<TMetadata>(
             this IMetadataProvider metadataProvider,
             Action<TMetadata> configureMetadata,
-            string metadataName = null,
-            bool ignoreCase = true,
-            bool searchInParent = true)
+            string metadataName = null)
             where TMetadata : new()
         {
             configureMetadata.AssertArgumentNotNull(nameof(configureMetadata));
 
-            metadataProvider.GetMetadataAsOption<TMetadata>(metadataName, ignoreCase, searchInParent)
+            metadataProvider.GetMetadataAsOption<TMetadata>(metadataName)
                 .Match(
                     some: configureMetadata,
                     none: () =>